(#217) context제거, useSyncExternalStore만을 사용한 커스텀 훅 제작
