+++ src/app/_components/clientProviders.tsx
@@ -2,7 +2,6 @@
 
 import { AuthProvider } from 'react-oidc-context'
 
-import { RemoteConfigProvider } from '@/context/remoteConfigContext'
 import { ToastProvider } from '@/context/toastContext'
 import { ChatAlertProvider } from '@/context/chatAlertContext'
 import { Toast } from '@/components/commons/toast'
@@ -15,18 +14,16 @@ import { SwrProvider } from './swrProvider'
 
 export function ClientProviders({ children }: { children: React.ReactNode }) {
   return (
-    <RemoteConfigProvider>
-      <AuthProvider userManager={userManager}>
-        <ToastProvider>
-          <ChatAlertProvider>
-            <SwrProvider>{children}</SwrProvider>
-            <Toast />
-            <ChatAlert />
-            <ServiceWorkerRegister />
-            <FcmListener />
-          </ChatAlertProvider>
-        </ToastProvider>
-      </AuthProvider>
-    </RemoteConfigProvider>
+    <AuthProvider userManager={userManager}>
+      <ToastProvider>
+        <ChatAlertProvider>
+          <SwrProvider>{children}</SwrProvider>
+          <Toast />
+          <ChatAlert />
+          <ServiceWorkerRegister />
+          <FcmListener />
+        </ChatAlertProvider>
+      </ToastProvider>
+    </AuthProvider>
   )
 }
